Cart

+++ MoustafaApp.Server/Dtos/CartDtos/CartItemDto.cs
@@ -6,7 +6,7 @@
         public int ProductId { get; set; }
         public string ProductName { get; set; } = null!;
         public int Quantity { get; set; }
-        public int PriceOfUnit { get; set; }
+        public decimal PriceOfUnit { get; set; }
         public string Photo { get; set; }= null!;
 
     }
